Refactor benchmarks and examples for code quality
- Update benchmark configurations and formatting
- Improve mesh reconstruction example
- Minor fixes in occupancy_gpu and mesh layers

+++ benches/tier1_optimizations.rs
@@ -3,7 +3,10 @@
 //! Tests BMI2, cache optimizations, prefetching, and specialized kernels
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
+
+#[cfg(target_arch = "x86_64")]
 use octaindex3d::morton;
+
 use octaindex3d::Route64;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
@@ -33,14 +36,17 @@ fn generate_routes(count: usize, seed: u64) -> Vec<Route64> {
         .collect()
 }
 
+
+
 /// Generate test coordinates for Morton encoding benchmarks
+#[cfg(target_arch = "x86_64")]
 fn generate_coords(count: usize, seed: u64) -> Vec<(u16, u16, u16)> {
     let mut rng = StdRng::seed_from_u64(seed);
     (0..count)
         .map(|_| {
-            let x = ((rng.random::<u16>() % 10000) * 2) as u16;
-            let y = ((rng.random::<u16>() % 10000) * 2) as u16;
-            let z = ((rng.random::<u16>() % 10000) * 2) as u16;
+            let x = (rng.random::<u16>() % 10000) * 2;
+            let y = (rng.random::<u16>() % 10000) * 2;
+            let z = (rng.random::<u16>() % 10000) * 2;
             (x, y, z)
         })
         .collect()
